Improved some examples, added better example for go channels

+++ 05-Monitoriai-cpp-cs-openmp/fork_join/main.cpp
@@ -9,18 +9,20 @@ using namespace std;
 using namespace std::chrono;
 
 bool is_mandelbrot(complex<double> number, int iterations);
+
 double compute_row(int col);
+
 double compute_column(int row);
 
 const double WIDTH = 2.5;
 const double HEIGHT = 2.5;
-const int ROWS = 2000;
-const int COLS = 2000;
+const size_t ROWS = 2000;
+const size_t COLS = 2000;
 
 // generates an image of Mandelbrot set
 int main() {
     // buffer for the image data. Each pixel is 3 bytes (red, green, blue)
-    auto* png_buffer = new uint8_t[ROWS * COLS * 3];
+    auto *png_buffer = new uint8_t[ROWS * COLS * 3];
     auto start = high_resolution_clock::now();  // start time measurement
 #pragma omp parallel for default(none) shared(png_buffer, COLS)
     // iterate over all columns and rows
@@ -40,17 +42,17 @@ int main() {
     auto stop = high_resolution_clock::now();
     // output image
     ofstream out("mandelbrot.png", std::ios::binary);
-    TinyPngOut png_out(static_cast<uint32_t>(COLS), static_cast<uint32_t>(ROWS), out);
-    png_out.write(png_buffer, static_cast<size_t>(COLS * ROWS));
+    TinyPngOut png_out(COLS, ROWS, out);
+    png_out.write(png_buffer, COLS * ROWS);
     // release memory
-    delete [] png_buffer;
+    delete[] png_buffer;
     auto duration = duration_cast<milliseconds>(stop - start);
     cout << "Image generated in " << duration.count() << " ms" << endl;
 }
 
 double compute_row(int col) {
     const complex<double> CENTER(-0.75, 0.0);
-    return CENTER.real() - WIDTH / 2.0 + (double)col * WIDTH / (double)COLS;
+    return CENTER.real() - WIDTH / 2.0 + (double) col * WIDTH / (double) COLS;
 }
 
 bool is_mandelbrot(complex<double> number, int iterations) {
@@ -65,5 +67,5 @@ bool is_mandelbrot(complex<double> number, int iterations) {
 
 double compute_column(int row) {
     const complex<double> CENTER(-0.75, 0.0);
-    return CENTER.imag() - HEIGHT / 2.0 + (double)row * HEIGHT / (double)ROWS;
+    return CENTER.imag() - HEIGHT / 2.0 + (double) row * HEIGHT / (double) ROWS;
 }
